fix: provision pending user invitation on sign
